feat: enhance warehouse management with input validation and improved error handling

+++ InventoryService/src/InventoryService.Application/DTOs/WarehouseUpdateRequest.cs
@@ -11,7 +11,7 @@ namespace InventoryService.Application.DTOs
         public string? Name { get; set; }
         public string? Location { get; set; }
         public Guid? ParentId { get; set; }
-        public decimal? Capacity { get; set; }
+        public int? Capacity { get; set; }
         public string? Status { get; set; }
         public bool? IsDeleted { get; set; }
     }
